requirements.txt
# Project Synapse - Windows Compatible Requirements
# Core LangGraph and AI Dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0

# LLM Providers
groq>=0.9.0
openai>=1.35.0
anthropic>=0.25.0

# Data Processing & Validation
pydantic>=2.7.0
typing-extensions>=4.12.0
python-dotenv>=1.0.0

# CLI & UI
typer>=0.12.0
rich>=13.7.0
click>=8.1.0

# HTTP & APIs
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.9.0

# Date & Time
python-dateutil>=2.9.0
pytz>=2024.1

# Monitoring & Logging
structlog>=24.1.0
loguru>=0.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Development Tools
black>=24.4.0
flake8>=7.0.0

# Security
cryptography>=42.0.0

# NOTE: uvloop removed - it doesn't support Windows
# asyncio (built-in) will be used instead
